requirements.txt
aiolimiter==1.2.1
anyio==4.4.0
APScheduler==3.10.4
async-lru==2.0.4
certifi==2024.6.2
contourpy==1.2.1
coverage==7.6.0
cycler==0.12.1
exceptiongroup==1.2.1
Faker==26.0.0
fonttools==4.53.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.9.0
mirakuru==2.5.2
mypy==1.11.2
mypy-extensions==1.0.0
numpy==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
polib==1.2.0
port-for==0.7.2
psutil==6.0.0
psycopg==3.2.1
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.6.1
pyparsing==3.1.2
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-postgresql==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.1
rsa==4.9
# Install from repo repo using RO token
git+https://github_pat_token_here@github.com/dev-rubik-love/backend_core.git@dev#egg=rubik_core
six==1.16.0
sniffio==1.3.1
Telethon==1.38.1
tomli==2.0.1
types-docutils==0.21.0.20240724
types-polib==1.2.0.20240811
types-psycopg2==2.9.21.20240819
types-Pygments==2.18.0.20240506
types-setuptools==74.0.0.20240831
typing_extensions==4.12.2
tzlocal==5.2
